fix: AI利用量カウンターの桁あふれ防止

+++ supabase/migrations/20260722050000_workspace_ai_nudge_controls.sql
@@ -1,10 +1,10 @@
 ALTER TABLE "workspaces"
   ADD COLUMN "ai_nudges_phase_one_enabled" boolean DEFAULT true NOT NULL,
   ADD COLUMN "ai_nudges_phase_two_enabled" boolean DEFAULT false NOT NULL,
-  ADD COLUMN "ai_nudges_phase_two_input_tokens" integer DEFAULT 0 NOT NULL,
-  ADD COLUMN "ai_nudges_phase_two_output_tokens" integer DEFAULT 0 NOT NULL,
-  ADD COLUMN "ai_nudges_phase_two_total_tokens" integer DEFAULT 0 NOT NULL,
-  ADD COLUMN "ai_nudges_phase_two_request_count" integer DEFAULT 0 NOT NULL;
+  ADD COLUMN "ai_nudges_phase_two_input_tokens" bigint DEFAULT 0 NOT NULL,
+  ADD COLUMN "ai_nudges_phase_two_output_tokens" bigint DEFAULT 0 NOT NULL,
+  ADD COLUMN "ai_nudges_phase_two_total_tokens" bigint DEFAULT 0 NOT NULL,
+  ADD COLUMN "ai_nudges_phase_two_request_count" bigint DEFAULT 0 NOT NULL;
 
 -- 既存ワークスペースは Phase 2 を既定OFFにするため、すでに表示中の LLM ナッジも
 -- 同時に隠す。再度ONにした後は、heartbeat が条件継続を確認したものだけ再表示する。
